refactor: rename Extra Drawer to Side Sheet
