+++ resources/assets/css/partials/vars.pcss
@@ -21,7 +21,7 @@
 
   --header-height: auto;
   --footer-height: 84px;
-  --extra-drawer-width: 25rem;
+  --side-sheet-width: 25rem;
   --sidebar-width: 20rem;
 
   --color-love: #bf2043;
@@ -32,6 +32,6 @@
   @media screen and (max-width: 768px) {
     --header-height: 56px;
     --footer-height: 96px;
-    --extra-drawer-width: 100%;
+    --side-sheet-width: 100%;
   }
 }
