+++ resources/assets/js/components/layout/main-wrapper/index.vue
@@ -2,7 +2,7 @@
   <div class="relative flex flex-1 overflow-hidden">
     <SideBar />
     <MainContent />
-    <ExtraDrawer />
+    <SideSheet />
     <ModalWrapper />
   </div>
 </template>
@@ -10,9 +10,9 @@
 <script lang="ts" setup>
 import { defineAsyncComponent } from 'vue'
 
-import SideBar from '@/components/layout/main-wrapper/sidebar/Sidebar.vue'
 import MainContent from '@/components/layout/main-wrapper/MainContent.vue'
-import ExtraDrawer from '@/components/layout/main-wrapper/extra-drawer/ExtraDrawer.vue'
+import SideBar from '@/components/layout/main-wrapper/sidebar/Sidebar.vue'
+import SideSheet from '@/components/layout/main-wrapper/side-sheet/SideSheet.vue'
 
 const ModalWrapper = defineAsyncComponent(() => import('@/components/layout/ModalWrapper.vue'))
 </script>
