+++ resources/assets/js/components/layout/main-wrapper/side-sheet/LogoutButton.vue
@@ -1,14 +1,14 @@
 <template>
-  <ExtraDrawerButton v-koel-tooltip.left title="Log out" @click.prevent="logout">
+  <SideSheetButton v-koel-tooltip.left title="Log out" @click.prevent="logout">
     <Icon :icon="faArrowRightFromBracket" />
-  </ExtraDrawerButton>
+  </SideSheetButton>
 </template>
 
 <script lang="ts" setup>
 import { faArrowRightFromBracket } from '@fortawesome/free-solid-svg-icons'
 import { eventBus } from '@/utils'
 
-import ExtraDrawerButton from '@/components/layout/main-wrapper/extra-drawer/ExtraDrawerButton.vue'
+import SideSheetButton from '@/components/layout/main-wrapper/side-sheet/SideSheetButton.vue'
 
 const logout = () => eventBus.emit('LOG_OUT')
 </script>
