+++ resources/assets/js/components/ui/SidebarMenuToggleButton.vue
@@ -1,8 +1,8 @@
 <template>
-  <ExtraDrawerButton class="block md:hidden" @click.prevent="toggleSidebar">
+  <SideSheetButton class="block md:hidden" @click.prevent="toggleSidebar">
     <Icon v-if="sidebarExpanded" :icon="faTimes" fixed-width />
     <Icon v-else :icon="faBars" fixed-width />
-  </ExtraDrawerButton>
+  </SideSheetButton>
 </template>
 
 <script lang="ts" setup>
@@ -10,7 +10,7 @@ import { faBars, faTimes } from '@fortawesome/free-solid-svg-icons'
 import { ref } from 'vue'
 import { eventBus } from '@/utils'
 
-import ExtraDrawerButton from '@/components/layout/main-wrapper/extra-drawer/ExtraDrawerButton.vue'
+import SideSheetButton from '@/components/layout/main-wrapper/side-sheet/SideSheetButton.vue'
 
 const sidebarExpanded = ref(false)
